# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth",
        "version": "1.0 2025-10-10"
    },
    "servers": [
        {
            "url": "/easy-auth/public"
        }
    ],
    "paths": {
...
        "/api/v1/user/change-password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Cambiar contraseña",
                "description": "Permite al usuario autenticado cambiar su contraseña",
                "operationId": "put-change-password",
                "requestBody": {
                    "description": "Datos para cambio de contraseña",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "old_password": {
                                        "description": "Contraseña actual del usuario",
                                        "type": "string",
                                        "format": "password",
                                        "example": "MiPasswordActual123!"
                                    },
                                    "new_password": {
                                        "description": "Nueva contraseña (mín. 8 caracteres, debe incluir mayúsculas, minúsculas, números y símbolos)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "MiNuevaPassword456@"
                                    },
                                    "new_password_confirmation": {
                                        "description": "Confirmación de la nueva contraseña",
                                        "type": "string",
                                        "format": "password",
                                        "example": "MiNuevaPassword456@"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambiada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contraseña actualizada exitosamente"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Contraseña actual incorrecta o no autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contraseña actual incorrecta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos proporcionados no son válidos"
                                        },
                                        "errors": {
                                            "properties": {
                                                "new_password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "La nueva contraseña debe tener al menos 8 caracteres."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/verification/code": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Validar el código de verificación y obtener el token",
                "description": "Valida un código de verificación de email. Devuelve un token de un solo uso si tiene éxito.",
                "operationId": "6e2f2eb9a1679b0818388154afdfcb3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Código verificado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "b2a1f3..."
                                                },
                                                "token_expires_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-10-22T04:00:00Z"
                                                },
                                                "identifier": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "email_verification"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Código incorrecto"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT obtenido del endpoint de login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}